chore: Split `TitleBarTemplateSettings` into partial files
