+++ src/Uno.UI/UI/Xaml/Controls/TitleBar/TitleBarTemplateSettings.mux.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+// MUX Reference TitleBarTemplateSettings.cpp, commit 5f9e85113
+
+namespace Microsoft.UI.Xaml.Controls;
+
+partial class TitleBarTemplateSettings
+{
+	public TitleBarTemplateSettings()
+	{
+		// EnsureProperties() is a generated method from TitleBarTemplateSettings.properties.cpp
+		// In Uno, dependency properties are registered via the static field initializers in TitleBarTemplateSettings.Properties.cs,
+		// so no explicit EnsureProperties() call is required.
+	}
+}
